+++ packages/ai/src/ui/chat-transport.ts
@@ -1,4 +1,4 @@
-import { UIMessageStreamPart } from '../ui-message-stream';
+import { UIMessageChunk } from '../ui-message-stream';
 import { ChatRequestOptions } from './chat';
 import { UIMessage } from './ui-messages';
 
@@ -15,11 +15,11 @@ export interface ChatTransport<UI_MESSAGE extends UIMessage> {
         | 'regenerate-assistant-message';
       messageId: string | undefined;
     } & ChatRequestOptions,
-  ) => Promise<ReadableStream<UIMessageStreamPart>>;
+  ) => Promise<ReadableStream<UIMessageChunk>>;
 
   reconnectToStream: (
     options: {
       chatId: string;
     } & ChatRequestOptions,
-  ) => Promise<ReadableStream<UIMessageStreamPart> | null>;
+  ) => Promise<ReadableStream<UIMessageChunk> | null>;
 }