+++ packages/anthropic/src/anthropic-messages-language-model.ts
@@ -339,11 +339,7 @@ export class AnthropicMessagesLanguageModel implements LanguageModelV2 {
         tool_choice: anthropicToolChoice,
       },
       warnings: [...warnings, ...toolWarnings],
-      betas: new Set([
-        'fine-grained-tool-streaming-2025-05-14',
-        ...messagesBetas,
-        ...toolsBetas,
-      ]),
+      betas: new Set([...messagesBetas, ...toolsBetas]),
       usesJsonResponseTool: jsonResponseTool != null,
     };
   }