+++ packages/rsc/tests/e2e/next-server/CHANGELOG.md
@@ -4,6 +4,39 @@
 
 ### Patch Changes
 
+- Updated dependencies [86cfc72]
+  - ai@5.0.0-beta.9
+
+## 0.0.1-beta.0
+
+### Patch Changes
+
+- Updated dependencies [6909543]
+- Updated dependencies [c8fce91]
+- Updated dependencies [9121250]
+  - ai@5.0.0-beta.8
+
+## 0.0.1-beta.0
+
+### Patch Changes
+
+- Updated dependencies [60132dd]
+  - ai@5.0.0-beta.7
+
+## 0.0.1-beta.0
+
+### Patch Changes
+
+- Updated dependencies [143c55b]
+- Updated dependencies [f04ffe4]
+- Updated dependencies [97c35c0]
+- Updated dependencies [fccf75c]
+  - ai@5.0.0-beta.6
+
+## 0.0.1-beta.0
+
+### Patch Changes
+
 - Updated dependencies [4f3e637]
   - ai@5.0.0-beta.5